requirements.txt
python-slugify~=8.0.4
icecream~=2.1.3
markdownify~=0.12.1
